+++ package.json
@@ -17,7 +17,6 @@
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.2",
     "react-scripts": "5.0.1",
-    "react-test-renderer": "^18.2.0",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-mock-store": "^1.5.4",
@@ -59,6 +58,7 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.33.0",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "react-test-renderer": "^18.2.0",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
